Force a full sync if active languages are newly added in project view
When there's no additional language specified in the project view file and a full sync is performed, the related targets are removed from the sync state and from the targetMap. Then, if a new language is added to the project view and an incremental sync is performed, the corresponding targets are not added back to the targetMap and the source files are shown unsynced (b/173528889).

This CL resolves this issue by forcing a full sync whenever a new language is added to the project view and a sync is requested. A more thorough solution would be to look into the ide-info files, find the targets matching with the newly added language and add them to the targetMap.

PiperOrigin-RevId: 446483072