Pool strings during BEP parsing
Adds an internal class to ParsedBepOutput for keeping track of the number of current parsing operations and strings representing information on seen file sets (uris, filenames, and path prefixes) that will later be parsed into OutputArtifacts.

When the BEP emits a NamedFileSet, it is replaced with a copy whose string references come from the shared pool, combining references for strings seen within a single parsing invocation and across concurrent parsing operations.

PiperOrigin-RevId: 449568789